ml/requirements.txt
audioread==2.1.8
cffi==1.14.0
decorator==4.4.2
joblib==0.14.1
librosa==0.7.2
llvmlite==0.31.0
numba==0.48.0
numpy==1.18.2
protobuf==3.11.3
pycparser==2.20
PyYAML==5.3.1
reformer-pytorch==0.19.7
resampy==0.2.2
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
SoundFile==0.10.3.post1
tensorboardX==2.0
termcolor==1.1.0
torch==1.4.0
tqdm==4.44.1
